+++ CHANGELOG.md
@@ -85,11 +85,13 @@ CHANGELOG
  * default ruleset: added avalanche rule.
  * new parameter `case_sensitive` (default: True)
 - Added wait expert for sleeping
+- Added domain suffix expert to extract the TLD/Suffix from a domain name.
 
 ### Harmonization
 - Renamed `JSON` to `JSONDict` and added a new type `JSON`. `JSONDict` saves data internally as JSON, but acts like a dictionary. `JSON` accepts any valid JSON.
 - fixed regex for `protocol.transport` it previously allowed more values than it should have.
 - New ASN type. Like integer but checks the range.
+- Added `(destination|source).domain_suffix` to hold the TLD/domain suffix.
 
 ### Requirements
 - Requests is no longer listed as dependency of the core. For depending bots the requirement is noted in their REQUIREMENTS.txt file